Added ProductManager and RecipeManager classes

+++ Class/Ingredient.cs
@@ -2,9 +2,9 @@ namespace Dietownik
 {
     class Ingredient : Product
     {
-        public double Weight { get; set; }
+        public decimal Weight { get; set; }
 
-        public Ingredient(string name, double weight) : base(name)
+        public Ingredient(string name, decimal weight) : base(name)
         {
             this.Weight = weight;
         }
